requirements.txt
mcp>=1.0.0
simple-salesforce>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0
